pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-registry"
version = "0.1.0"
description = "Version-controlled prompt templates with variables"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
]
keywords = ["prompt", "template", "llm", "jinja2", "registry"]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
prompt-registry = "prompt_registry.cli:main"

[project.urls]
Homepage = "https://github.com/cognitioncommons/prompt-registry"
Documentation = "https://github.com/cognitioncommons/prompt-registry#readme"
Repository = "https://github.com/cognitioncommons/prompt-registry.git"
Issues = "https://github.com/cognitioncommons/prompt-registry/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["prompt_registry*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=prompt_registry --cov-report=term-missing"
